+++ packages/expo-dev-launcher/android/src/with-updates/expo/modules/devlauncher/DevLauncherUpdatesInterfaceDelegate.kt
@@ -5,6 +5,6 @@ import expo.modules.updates.UpdatesController
 
 object DevLauncherUpdatesInterfaceDelegate {
   fun initializeUpdatesInterface(context: Context) {
-    DevLauncherController.instance.updatesInterface = UpdatesController.initializeAsDevLauncherWithoutStarting(context)
+    DevLauncherController.instance.updatesInterface = UpdatesController.initializeAsDevLauncherWithoutStarting(context, DevLauncherController.instance)
   }
 }