+++ packages/expo-updates-interface/android/src/main/java/expo/modules/updatesinterface/UpdatesInterfaceCallbacks.java
@@ -1,0 +1,5 @@
+package expo.modules.updatesinterface;
+
+public interface UpdatesInterfaceCallbacks {
+    void onRequestRelaunch();
+}