+++ packages/expo-updates-interface/ios/EXUpdatesInterface/UpdatesExternalInterface.swift
@@ -20,6 +20,7 @@ public typealias UpdatesManifestBlock = (_ manifest: [String: Any]) -> Bool
 @objc(EXUpdatesExternalInterface)
 public protocol UpdatesExternalInterface {
   @objc weak var bridge: AnyObject? { get set }
+  @objc weak var updatesExternalInterfaceDelegate: (any UpdatesExternalInterfaceDelegate)? { get set }
   @objc var launchAssetURL: URL? { get }
 
   @objc func reset()
@@ -31,4 +32,12 @@ public protocol UpdatesExternalInterface {
     success successBlock: @escaping UpdatesUpdateSuccessBlock,
     error errorBlock: @escaping UpdatesErrorBlock
   )
+}
+
+/**
+ * Protocol for communication/delegation back to the host dev client for functionality.
+ */
+@objc(EXUpdatesExternalInterfaceDelegate)
+public protocol UpdatesExternalInterfaceDelegate {
+  @objc func updatesExternalInterfaceDidRequestRelaunch(_ updatesExternalInterface: UpdatesExternalInterface)
 }