+++ packages/expo-updates/android/src/main/java/expo/modules/updates/UpdatesController.kt
@@ -6,6 +6,7 @@ import com.facebook.react.ReactNativeHost
 import expo.modules.updates.loader.LoaderTask
 import expo.modules.updates.logging.UpdatesErrorCode
 import expo.modules.updates.logging.UpdatesLogger
+import expo.modules.updatesinterface.UpdatesInterfaceCallbacks
 
 /**
  * Main entry point to expo-updates. Singleton that keeps track of updates state, holds references
@@ -71,7 +72,7 @@ class UpdatesController {
       }
     }
 
-    @JvmStatic fun initializeAsDevLauncherWithoutStarting(context: Context): UpdatesDevLauncherController {
+    @JvmStatic fun initializeAsDevLauncherWithoutStarting(context: Context, callbacks: UpdatesInterfaceCallbacks): UpdatesDevLauncherController {
       check(singletonInstance == null) { "UpdatesController must not be initialized prior to calling initializeAsDevLauncherWithoutStarting" }
 
       var updatesDirectoryException: Exception? = null
@@ -92,7 +93,8 @@ class UpdatesController {
         initialUpdatesConfiguration,
         updatesDirectory,
         updatesDirectoryException,
-        UpdatesConfiguration.isMissingRuntimeVersion(context, null)
+        UpdatesConfiguration.isMissingRuntimeVersion(context, null),
+        callbacks
       )
       singletonInstance = instance
       return instance